Update ManagedExecutorServiceImpl.java
Needs to return an Integer, not a long

+++ container/openejb-core/src/main/java/org/apache/openejb/threads/impl/ManagedExecutorServiceImpl.java
@@ -101,7 +101,7 @@ public class ManagedExecutorServiceImpl extends AbstractExecutorService implemen
         }
     }
 
-    public Long getQueueSize() {
+    public Integer getQueueSize() {
         if (delegate instanceof ThreadPoolExecutor) {
             return ((ThreadPoolExecutor) delegate).getQueue().size();
         } else {
